"Updated models, services, and view models to reflect changes in data types and added new functionality to navigate to transactions view."

+++ lib/models/cuenta_model.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 class CuentaModel {
   int id;
   int alumnoId;
-  int saldo;
+  double saldo;
   DateTime fechaApertura;
   bool solvente;
